test: use zaptest for logging
